DONATION, LOCATION, MAJLIS  Entity and Dto Created

+++ ProjectName.Infra/Repo/GenericRepo.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Http.HttpResults;
+using Microsoft.EntityFrameworkCore;
 using ProjectName.Domain.Base;
 using ProjectName.Infra.Context;
 using System.Linq.Expressions;
@@ -6,7 +7,7 @@ using X.PagedList;
 
 namespace ProjectName.Infra.Repo
 {
-    public class GenericRepo<T> : IGenericRepo<T> where T : class
+  public class GenericRepo<T> : IGenericRepo<T> where T : class
   {
     private readonly DBCntxt _context;
     private readonly DbSet<T> _db;
@@ -18,7 +19,7 @@ namespace ProjectName.Infra.Repo
     public async Task Delete(int id)
     {
       var entity = await _db.FindAsync(id);
-      _db.Remove(entity);
+      if (entity != null) _db.Remove(entity);
     }
 
     public void DeleteRange(IEnumerable<T> entities)
@@ -52,8 +53,8 @@ namespace ProjectName.Infra.Repo
           query = query.Include(item);
         }
       }
-
-      return await query.AsNoTracking().FirstOrDefaultAsync(expression);
+      var result = await query.AsNoTracking().FirstOrDefaultAsync(expression);
+      return result; //if (result != null) return result; //
     }
 
     public async Task<List<T>> Gets(
@@ -83,12 +84,12 @@ namespace ProjectName.Infra.Repo
     }
 
     public async Task<IPagedList<T>> Gets(
-      BaseDTOPagination req,
+      BaseDtoPagination req,
       List<string>? includes = null)
     {
       if (req == null)
       {
-        req = new BaseDTOPagination()
+        req = new BaseDtoPagination()
         {
           PageNo = 1,
           PageSize = 2
@@ -103,13 +104,12 @@ namespace ProjectName.Infra.Repo
         }
       }
       return await query.AsNoTracking()
-        .ToList().ToPagedListAsync(req.PageNo ?? 1, req.PageSize);
+        .ToPagedListAsync(req.PageNo ?? 1, req.PageSize);
     }
 
-    public async Task<IPagedList<T>> Gets<TDto>(PaginateRequestFilter<T, TDto?>? req) 
+    public async Task<IPagedList<T>> Gets<TDto>(PaginateRequestFilter<T, TDto>? req)
       where TDto : class
     {
-      Console.WriteLine("----------PaginateRequestFilter-------------");
       if (req == null)
       {
         req = new PaginateRequestFilter<T, TDto>()
@@ -124,99 +124,102 @@ namespace ProjectName.Infra.Repo
       IQueryable<T> query = _db;
       query = query.FilterByGeneric<T, TDto>(req.Search);
       query = query.OrderByGeneric<T>(req.Sort);
-
-      //if (req.includes != null)
-      //{
-      //  foreach (var item in includes)
-      //  {
-      //    query = query.Include(item);
-      //  }
-      //}
+      /*
+        if (query.includes != null)
+        {
+          foreach (var item in includes)
+          {
+            query = query.Include(item);
+          }
+        }
+      */
       return await query
         .AsNoTracking()
         .ToPagedListAsync(req.PageNo, req.PageSize);
     }
   }
 }
-// https://blog.zhaytam.com/2020/05/17/dynamic-sorting-filtering-csharp/
+/**
+ https://blog.zhaytam.com/2020/05/17/dynamic-sorting-filtering-csharp/
 
-// APPROACH 1. WITHOUT LIBRARY
-// MANUAL DYNAMIC SORTING
-//public IQueryable<Product> ApplyOrderBy(IQueryable<Product> query, string property)
-//{
-//  switch (property)
-//  {
-//    case "Name":
-//      return query.OrderBy(p => p.Name);
-//    case "Quantity":
-//      return query.OrderBy(p => p.Quantity);
-//    case "Price":
-//      return query.OrderBy(p => p.Price);
-//    default:
-//      return query;
-//  }
-//}
+APPROACH 1.WITHOUT LIBRARY
+MANUAL DYNAMIC SORTING
+public IQueryable<Product> ApplyOrderBy(IQueryable<Product> query, string property)
+{
+  switch (property)
+  {
+    case "Name":
+      return query.OrderBy(p => p.Name);
+    case "Quantity":
+      return query.OrderBy(p => p.Quantity);
+    case "Price":
+      return query.OrderBy(p => p.Price);
+    default:
+      return query;
+  }
+}
 
-//var query = _dbContext.Products.AsQueryable();
-//query = ApplyOrderBy(query, propertySentByUser);
+var query = _dbContext.Products.AsQueryable();
+query = ApplyOrderBy(query, propertySentByUser);
 
-// MANUAL DYNAMIC FILTERING
-//public IQueryable<Product> ApplyFilter(IQueryable<Product> query, string property, object value)
-//{
-//  switch (property)
-//  {
-//    case "Name":
-//      return query.Where(p => p.Name.Contains(value.ToString()));
-//    case "Quantity":
-//      return query.Where(p => p.Quantity >= value);
-//    case "Price":
-//      return query.Where(p => p.Price >= value);
-//    default:
-//      return query;
-//  }
-//}
-// MANUAL DYNAMIC ADVANCE FILTER
-//(Product.Brand == "Nike" || Product.Brand == "Adidas")
-//&& (Product.Price >= 20 && Product.Price <= 100)
-//&& Product.Enabled
-
-
-
-// APPROACH 2. WITH DYNAMIC.NET LIBRARY
-// Dynamic.NET ->  TODO
-
-// DYNAMIC SORTING
-//public async Task<IPagedList<T>> GetPagetListAlternate()
-//{
-//  var propertyGetter = DynamicExpressions.GetPropertyGetter<Product>(propertySentByUser);
-//  // ^ can be cached or even compiled to a Func<Product, object>
-//  var query = _dbContext.Products.AsQueryable().OrderBy(propertyGetter);
-//}
-
-// DYNAMIC FILTERING
-//var predicate = DynamicExpressions.GetPredicate<Product>(propertySentByUser, operatorSentByUser, valueSentByUser);
-//// ^ can also be cached or compiled and used anywhere
-//var products = _dbContext.Products.AsQueryable().Where(predicate).ToList();
-//// ^ or FirstByDefault, Any, etc...
-
-// DYNAMIC FILTERING LIBRARY APPROACH
-//var predicate = new DynamicFilterBuilder<Product>()
-//  .And("Enabled", FilterOperator.Equals, true)
-//  .And(b => b.And("Brand", FilterOperator.Equals, "Nike").Or("Brand", FilterOperator.Equals, "Adidas"))
-//  .And(b => b.And("Price", FilterOperator.GreaterThanOrEqual, 20).And("Price", FilterOperator.LessThanOrEqual, 100))
-//  .Build();
-
-//var products = _dbContext.Products.AsQueryable().Where(predicate).ToList()
+MANUAL DYNAMIC FILTERING
+public IQueryable<Product> ApplyFilter(IQueryable<Product> query, string property, object value)
+{
+  switch (property)
+  {
+    case "Name":
+      return query.Where(p => p.Name.Contains(value.ToString()));
+    case "Quantity":
+      return query.Where(p => p.Quantity >= value);
+    case "Price":
+      return query.Where(p => p.Price >= value);
+    default:
+      return query;
+  }
+}
+MANUAL DYNAMIC ADVANCE FILTER
+(Product.Brand == "Nike" || Product.Brand == "Adidas")
+&& (Product.Price >= 20 && Product.Price <= 100)
+&& Product.Enabled
 
 
-// HOW DOES IT WORKS UNDER THE HOOD
-//public static Expression<Func<TEntity, object>> GetPropertyGetter<TEntity>(string property)
-//{
-//  if (property == null)
-//    throw new ArgumentNullException(nameof(property));
 
-//  var param = Expression.Parameter(typeof(TEntity));
-//  var prop = param.GetNestedProperty(property);
-//  var convertedProp = Expression.Convert(prop, typeof(object));
-//  return Expression.Lambda<Func<TEntity, object>>(convertedProp, param);
-//}
+ APPROACH 2. WITH DYNAMIC.NET LIBRARY
+ Dynamic.NET ->  TODO
+
+ DYNAMIC SORTING
+public async Task<IPagedList<T>> GetPagetListAlternate()
+{
+  var propertyGetter = DynamicExpressions.GetPropertyGetter<Product>(propertySentByUser);
+  // ^ can be cached or even compiled to a Func<Product, object>
+  var query = _dbContext.Products.AsQueryable().OrderBy(propertyGetter);
+}
+
+ DYNAMIC FILTERING
+var predicate = DynamicExpressions.GetPredicate<Product>(propertySentByUser, operatorSentByUser, valueSentByUser);
+// ^ can also be cached or compiled and used anywhere
+var products = _dbContext.Products.AsQueryable().Where(predicate).ToList();
+// ^ or FirstByDefault, Any, etc...
+
+ DYNAMIC FILTERING LIBRARY APPROACH
+var predicate = new DynamicFilterBuilder<Product>()
+  .And("Enabled", FilterOperator.Equals, true)
+  .And(b => b.And("Brand", FilterOperator.Equals, "Nike").Or("Brand", FilterOperator.Equals, "Adidas"))
+  .And(b => b.And("Price", FilterOperator.GreaterThanOrEqual, 20).And("Price", FilterOperator.LessThanOrEqual, 100))
+  .Build();
+
+var products = _dbContext.Products.AsQueryable().Where(predicate).ToList()
+
+
+ HOW DOES IT WORKS UNDER THE HOOD
+public static Expression<Func<TEntity, object>> GetPropertyGetter<TEntity>(string property)
+{
+  if (property == null)
+    throw new ArgumentNullException(nameof(property));
+
+  var param = Expression.Parameter(typeof(TEntity));
+  var prop = param.GetNestedProperty(property);
+  var convertedProp = Expression.Convert(prop, typeof(object));
+  return Expression.Lambda<Func<TEntity, object>>(convertedProp, param);
+}
+*/
